+++ SRC/unsigned_format.c
@@ -6,13 +6,13 @@
 /*   By: lcoissar <lcoissar@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 10:41:46 by lcoissar          #+#    #+#             */
-/*   Updated: 2022/11/19 07:58:20 by lcoissar         ###   ########lyon.fr   */
+/*   Updated: 2022/11/21 08:21:14 by lcoissar         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../INC/ft_printf.h"
 
-int	unsigned_format(unsigned int ul)
+size_t	unsigned_format(unsigned int ul)
 {
 	char	*str;
 	int		len;
